Added Expected Discounted Payoff computation.
In gt-to-pomdp.py, made additional CMD args optional by adding '-' to beginning of tag.

In models.py, gave each player its own payoff function - this must be filled in by something external (like GTModel or PseudoPOMDPModel). Removed probability_lookup function, since it is not longer used. Added Expected Discounted Payoff function to PseudoPOMDPModel - it is exposed as a property, so it only calculates the payoff when it needs to. Changed states, observations, and actions to lists instead of sets - sets are unordered, so the output could change on different runs. We also don't check for set inclusion on these sets, so we don't need the extra speed.
Gave POMDPModel the Expected Discounted Payoff as well, but it is just a copy of the PseudoPOMDPModel's Expected Discounted Payoff.