Fixed matrix creation in Value Function generation.
Some matricies were being calculated incorrectly - this worked for the simple case because the number of actions were equal to the number of FSA states, and every FSA state mapped to a different action. In general, however, it was wrong. They are now calculated correctly.

Also fixed how Pi was generated to work in general, instead of assuming  2x2 sub-matricies.