Added belief_guided as a post-pruning method for V and Q type algorithms (e.g. enum, incremental pruning), and as a regular pruning method for incremental pruning.
This method can be activated by passing the flag -enum_purge belief_guided or -q_purge belief_guided for post-pruning, and -inc_prune belief_guided for incremental pruning. When using belief_guided, an additional argument -belief_states must be passed as well, and will be used as the initial points to generate from.